Migrate expression string tests to correct test directory
